chore: split ops and derivation
