+++ static/src/utils/shared.ts
@@ -1,0 +1,4 @@
+// Helper function to convert a string to Uint8Array
+function strToUint8Array(str: string): Uint8Array {
+    return new TextEncoder().encode(str);
+}
