+++ static/src/utils/strings.ts
@@ -1,0 +1,1 @@
+export const PUBLIC_KEY = "public-key"
